show edit/link state visually to user
This lets the user see their entry, as well as which links have been
set by toggling transparency values for the edit link buttons.
It also shows a pointer for the title link which unfortunately
surfaces the fact that the entire title block is the button for
opening a link as opposed to just the title text as the user will
likely expect. So next will be trying to use the .titleContent as
a wrapper for a span and hopefully not having to write a line by
line title builder.