add alarms and re-implement background page
After some searching it looks like webRequests don't work with
event pages, and the event page equivalent "declarativeWebRequests"
API appears to have died in its cradle in the chrome beta channel.
So despite the warnings that existing background pages should be
migrated to event pages, some of the critical active syncing
functionality is only easily done with background pages. It could
also be done with websockets, but that would effectively be rewriting
webRequests from scratch and would really defeat the purpose of
making this a browser extension in the first place. Therefore, I'm
sticking with background pages until forced to do otherwise.

Also in this commit is the ability to make the program aware it is the first
run of the extension, some renaming to clarify a couple function names
to clean up the global namespace, and the addition of alarm declarers
and listeners for passive syncing.

I am still not sure if releases should have all their latest release checked
automatically on the first install. This is an expensive process for both
the user and MU, and might be better left to the user to have to check each
one manually.