add options menu composer
The popup builder script is heavy enough so I fanned this out to a new file.
I'm not necessarily proud of this mega-switch-case data infrastructure but it
didn't seem worthwhile to implement something more scalable when presumably
there won't be that many options total. A couple of the options already seem
kind of unnecessary and might get pruned. Next step is to hook up the css
and implement the event handlers for the menu.