add review stage to mark up-to-date useflow
The user clicks once to update the series, then again to send it down. This
gives the user a chance to review the new chapter has been updated as well as
make changes or click its link. Clicking the button again sends it down to its
alphabetically sorted position. This required yet another flag DOM attribute
on the up-to-date button to remove it from the sort conditions, but this
should be it for adding functionality to it. Some code was also reworked to
make it more generalized but it's doubtful any of it will be useful for
any future features, so it's mostly aesthetics and housekeeping.