add 2-click up-to-date and initial release check
There is now the functionality to check for releases for series which have
not yet gotten a new release. This might be harmful to user's understanding
of how the application works, though. They might expect to be able to use
it to check for a release any time rather than just after first adding
the series to their list. A better prompt description and giving a visual
indication that the latest release has not yet been initialized would help to
mitigate this.

The up-to-date button now has 3 separate functions: check for a new release,
mark the release up to date, and sort it back down alphabetically amongst the
rest of the series with no new releases. Which function is available is
determined by the context. In case this is perceived as inconvenient or
confusing for the user, an option to combine all three functions into one
click of the 'Mark Up-to-Date' button is available. My preference is the
2-3 click button because it offers more control, but it might be that the
1-click option would be better as the default for the users.