delay disappearing scrollbar
The newest effort in my war against the scrollbar. Previously, if there weren't
enough series rows (or other relatively positioned elements inside the popup
class) to fill the popup to make it expand beyond the height of its container,
scrolling wouldn't work. So this hack leaves a pixel at the bottom of the popup
overflowing its height until the browser has finished laying everything out.
This kind of hack is most likely browser-specific, so it's good that extensions
only target a specific browser. It would be better to find a cleaner workaround
though.