centralize popup list state updates
This commit is larger than it really should be because I kept spotting
inefficiencies and places where async functions were misused. Now when
there are changes to series that could affect anything the user needs
to see or know, the lists involved are updated by a central list state
update function. This changes their option text to determine how many
unread releases are in them, and if its list table is loaded in the DOM
it just unloads them to defer calculations and avoid DOM traversals.
The alternative is piecewise updating anything that has changed which
is really unnecessary.

Changes to the current list are still mainly done by the functions
doing the changes, and the list state update is called at the end to
handle anything unnecessary to immediately showing the user what is
happening. The list state update also calls the badge text to update,
which is done by the background script. This was partially done because
the debugger was throwing odd errors at me whenever the badge text
was changed and I wanted it out of my way. But since the background
script is also already doing updates and syncs and updating the
badge text there, and since the badge text isn't actually an element
of the popup, it seemed like a good fit.

Lastly there was some refactoring done to the popup builder select
dropdown and option building to make it cleaner to call an update
function on current list options.