test full data pull
This does a full go through the list and series pulling mill, stubbing
the 5 XHR functions with test data in their place. The test needs all the
lists to be primed ahead of time, an annoying gotcha when I was trying
to make the functions more modular and was forced to pass a boolean for
prime/don't prime to make it work all the way through the chain.