add fake random popup cli option
Running ktest.js now has the option to create a fake popup to manually
debug issues with a random data set. Only the first list loads correctly,
it will be necessary to stub out a large number of chrome extension api
calls to use it the same way as the popup.