Add release tab feature
Large commit as it took awhile to get to an unbroken build. The main thing added
is a release tab which currently only allows you to view releases. The "paging"
functionality, which I think was one of the last things added, was too rigid
and several functions needed reworked to accommodate having a release page
which works similarly but does not have the same DOM elements as the series
page. Also went through and added function headers where there were none.

The things broken in this commit include: navigating to options page (which is
now inconsistent and should be integrated with the page system), filter does
not work for releases, performance is terrible, new buttons are poorly
integrated with style.